refactor to have only one abstraction: source ingesters
