+++ ee/maintained-apps/inputs/homebrew/1password.json
@@ -1,7 +1,7 @@
 {
   "name": "1Password",
   "unique_identifier": "com.1password.1password",
-  "source_identifier": "1password",
+  "token": "1password",
   "installer_format": "zip",
-  "source": "homebrew"
+  "slug": "1password/darwin"
 }
