+++ ee/maintained-apps/inputs/winget/notion.json
@@ -1,0 +1,4 @@
+{
+  "name": "Notion",
+  "slug": "notion/windows"
+}
